requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.1
cryptography==42.0.5
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
rich==13.7.0
python-dateutil==2.9.0
